+++ boot/freeldr/freeldr/arch/arm/macharm.c
@@ -95,19 +95,6 @@ ArmPrepareForReactOS(VOID)
     return;
 }
 
-BOOLEAN
-ArmDiskGetBootPath(OUT PCHAR BootPath, IN ULONG Size)
-{
-    PCCH Path = "ramdisk(0)";
-
-    /* Make sure enough space exists */
-    if (Size < sizeof(Path)) return FALSE;
-
-    /* On ARM platforms, the loader is always in RAM */
-    strcpy(BootPath, Path);
-    return TRUE;
-}
-
 PCONFIGURATION_COMPONENT_DATA
 ArmHwDetect(VOID)
 {
@@ -152,7 +139,12 @@ BOOLEAN
 ArmInitializeBootDevices(VOID)
 {
     /* Emulate old behavior */
-    return (ArmHwDetect() != NULL);
+    if (ArmHwDetect() == NULL)
+        return FALSE;
+
+    /* On ARM platforms, the loader is always in RAM */
+    strcpy(FrldrBootPath, "ramdisk(0)");
+    return TRUE;
 }
 
 FREELDR_MEMORY_DESCRIPTOR ArmMemoryMap[32];
@@ -239,6 +231,5 @@ MachInit(IN PCCH CommandLine)
     MachVtbl.GetMemoryMap = ArmMemGetMemoryMap;
     MachVtbl.InitializeBootDevices = ArmInitializeBootDevices;
     MachVtbl.HwDetect = ArmHwDetect;
-    MachVtbl.DiskGetBootPath = ArmDiskGetBootPath;
     MachVtbl.HwIdle = ArmHwIdle;
 }
