+++ boot/freeldr/freeldr/arch/i386/entry.S
@@ -62,7 +62,6 @@ PUBLIC ContinueAddress
 ContinueAddress:
     .long _FrldrStartup
 
-
 _FrldrStartup:
 
     /* Store BootDrive and BootPartition */
@@ -98,6 +97,47 @@ stop:
     jmp short stop
     nop
     nop
+
+
+PUBLIC _Reboot
+_Reboot:
+    /* Set the function ID */
+    mov bx, FNID_Reboot
+
+    /* Switch to real mode (we don't return) */
+    jmp SwitchToReal
+
+
+/*
+ * VOID __cdecl ChainLoadBiosBootSectorCode(
+ *     IN UCHAR BootDrive OPTIONAL,
+ *     IN ULONG BootPartition OPTIONAL);
+ *
+ * RETURNS: Nothing
+ */
+PUBLIC _ChainLoadBiosBootSectorCode
+_ChainLoadBiosBootSectorCode:
+    /* Set the boot drive */
+    mov dl, [esp + 4]
+    test dl, dl
+    jnz set_part
+    mov dl, byte ptr ds:[_FrldrBootDrive]
+
+    /* Set the boot partition */
+set_part:
+    mov eax, [esp + 8]
+    test eax, eax
+    jnz continue
+    mov eax, dword ptr ds:[_FrldrBootPartition]
+continue:
+    /* Store the 1-byte truncated partition number in DH */
+    mov dh, al
+
+    /* Set the function ID */
+    mov bx, FNID_ChainLoadBiosBootSectorCode
+
+    /* Switch to real mode (we don't return) */
+    jmp SwitchToReal
 
 
 /*
@@ -137,27 +177,6 @@ _PxeCallApi:
     mov ax, word ptr [BSS_PxeResult]
 
     ret
-
-
-PUBLIC _Reboot
-_Reboot:
-    /* Set the function ID */
-    mov bx, FNID_Reboot
-
-    /* Switch to real mode (we don't return) */
-    jmp SwitchToReal
-
-
-PUBLIC _ChainLoadBiosBootSectorCode
-_ChainLoadBiosBootSectorCode:
-    /* Set the boot drive */
-    mov dl, byte ptr ds:[_FrldrBootDrive]
-
-    /* Set the function ID */
-    mov bx, FNID_ChainLoadBiosBootSectorCode
-
-    /* Switch to real mode (we don't return) */
-    jmp SwitchToReal
 
 
 PUBLIC i386CallRealMode
