+++ boot/freeldr/freeldr/arch/i386/pcdisk.c
@@ -73,6 +73,65 @@ typedef struct
 
 #include <poppack.h>
 
+/* DISK IO ERROR SUPPORT *****************************************************/
+
+static BOOLEAN bReportError = TRUE;
+
+VOID DiskReportError(BOOLEAN bError)
+{
+    bReportError = bError;
+}
+
+static PCSTR DiskGetErrorCodeString(ULONG ErrorCode)
+{
+    switch (ErrorCode)
+    {
+    case 0x00:  return "no error";
+    case 0x01:  return "bad command passed to driver";
+    case 0x02:  return "address mark not found or bad sector";
+    case 0x03:  return "diskette write protect error";
+    case 0x04:  return "sector not found";
+    case 0x05:  return "fixed disk reset failed";
+    case 0x06:  return "diskette changed or removed";
+    case 0x07:  return "bad fixed disk parameter table";
+    case 0x08:  return "DMA overrun";
+    case 0x09:  return "DMA access across 64k boundary";
+    case 0x0A:  return "bad fixed disk sector flag";
+    case 0x0B:  return "bad fixed disk cylinder";
+    case 0x0C:  return "unsupported track/invalid media";
+    case 0x0D:  return "invalid number of sectors on fixed disk format";
+    case 0x0E:  return "fixed disk controlled data address mark detected";
+    case 0x0F:  return "fixed disk DMA arbitration level out of range";
+    case 0x10:  return "ECC/CRC error on disk read";
+    case 0x11:  return "recoverable fixed disk data error, data fixed by ECC";
+    case 0x20:  return "controller error (NEC for floppies)";
+    case 0x40:  return "seek failure";
+    case 0x80:  return "time out, drive not ready";
+    case 0xAA:  return "fixed disk drive not ready";
+    case 0xBB:  return "fixed disk undefined error";
+    case 0xCC:  return "fixed disk write fault on selected drive";
+    case 0xE0:  return "fixed disk status error/Error reg = 0";
+    case 0xFF:  return "sense operation failed";
+
+    default:    return "unknown error code";
+    }
+}
+
+static VOID DiskError(PCSTR ErrorString, ULONG ErrorCode)
+{
+    CHAR ErrorCodeString[200];
+
+    if (bReportError == FALSE)
+        return;
+
+    sprintf(ErrorCodeString, "%s\n\nError Code: 0x%lx\nError: %s",
+            ErrorString, ErrorCode, DiskGetErrorCodeString(ErrorCode));
+
+    TRACE("%s\n", ErrorCodeString);
+
+    UiMessageBox(ErrorCodeString);
+}
+
 /* FUNCTIONS *****************************************************************/
 
 BOOLEAN DiskResetController(UCHAR DriveNumber)
@@ -157,7 +216,10 @@ static BOOLEAN PcDiskReadLogicalSectorsLBA(UCHAR DriveNumber, ULONGLONG SectorNu
     }
 
     /* If we get here then the read failed */
-    ERR("Disk Read Failed in LBA mode: %x (DriveNumber: 0x%x SectorNumber: %I64d SectorCount: %d)\n", RegsOut.b.ah, DriveNumber, SectorNumber, SectorCount);
+    DiskError("Disk Read Failed in LBA mode", RegsOut.b.ah);
+    ERR("Disk Read Failed in LBA mode: %x (%s) (DriveNumber: 0x%x SectorNumber: %I64d SectorCount: %d)\n",
+        RegsOut.b.ah, DiskGetErrorCodeString(RegsOut.b.ah),
+        DriveNumber, SectorNumber, SectorCount);
 
     return FALSE;
 }
@@ -271,7 +333,10 @@ static BOOLEAN PcDiskReadLogicalSectorsCHS(UCHAR DriveNumber, ULONGLONG SectorNu
         /* If we retried 3 times then fail */
         if (RetryCount >= 3)
         {
-            ERR("Disk Read Failed in CHS mode, after retrying 3 times: %x\n", RegsOut.b.ah);
+            DiskError("Disk Read Failed in CHS mode, after retrying 3 times", RegsOut.b.ah);
+            ERR("Disk Read Failed in CHS mode, after retrying 3 times: %x (%s) (DriveNumber: 0x%x SectorNumber: %I64d SectorCount: %d)\n",
+                RegsOut.b.ah, DiskGetErrorCodeString(RegsOut.b.ah),
+                DriveNumber, SectorNumber, SectorCount);
             return FALSE;
         }
 
@@ -563,7 +628,7 @@ PcDiskGetCacheableBlockCount(UCHAR DriveNumber)
     }
     /* Get the disk geometry. If this fails then we will
      * just return 1 sector to be safe. */
-    else if (! PcDiskGetDriveGeometry(DriveNumber, &Geometry))
+    else if (!PcDiskGetDriveGeometry(DriveNumber, &Geometry))
     {
         return 1;
     }
@@ -571,35 +636,6 @@ PcDiskGetCacheableBlockCount(UCHAR DriveNumber)
     {
         return Geometry.Sectors;
     }
-}
-
-BOOLEAN
-PcDiskGetBootPath(OUT PCHAR BootPath, IN ULONG Size)
-{
-    // FIXME: Keep it there, or put it in DiskGetBootPath?
-    // Or, abstract the notion of network booting to make
-    // sense for other platforms than the PC (and this idea
-    // already exists), then we would need to check whether
-    // we were booting from network (and: PC --> PXE, etc...)
-    // and if so, set the correct ARC path. But then this new
-    // logic could be moved back to DiskGetBootPath...
-
-    if (*FrldrBootPath)
-    {
-        /* Copy back the buffer */
-        if (Size < strlen(FrldrBootPath) + 1)
-            return FALSE;
-        strncpy(BootPath, FrldrBootPath, Size);
-        return TRUE;
-    }
-
-    // FIXME! FIXME! Do this in some drive recognition procedure!!!!
-    if (PxeInit())
-    {
-        strcpy(BootPath, "net(0)");
-        return TRUE;
-    }
-    return DiskGetBootPath(BootPath, Size);
 }
 
 /* EOF */
