+++ boot/freeldr/freeldr/arch/realmode/amd64.S
@@ -273,7 +273,7 @@ InRealMode:
     /* Restore real mode stack */
     mov sp, word ptr ds:[stack16]
 
-   // sti /* These are ok now */
+    // sti /* These are ok now */
 
     /* Do the callback, specified by bx */
     shl bx, 1
